+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## 0.1.3 (2025-09-02)
 - Add some attributes default arguments to HtmlContainer addHeader()
+- Add <i> and <b> as HtmlTag variants
 
 ## 0.1.2 (2025-08-30)
 - Add more methods to HtmlPage()